+++ src/Utils/BufferNodeUtils.js
@@ -1,5 +1,3 @@
-import { setTriangle } from './TriangleUtils.js';
-
 export function arrayToBox( nodeIndex32, array, target ) {
 
 	target.min.x = array[ nodeIndex32 ];
@@ -9,6 +7,38 @@ export function arrayToBox( nodeIndex32, array, target ) {
 	target.max.x = array[ nodeIndex32 + 3 ];
 	target.max.y = array[ nodeIndex32 + 4 ];
 	target.max.z = array[ nodeIndex32 + 5 ];
+
+}
+
+// sets the vertices of triangle `tri` with the 3 vertices after i
+export function setTriangle( tri, i, index, pos ) {
+
+	const ta = tri.a;
+	const tb = tri.b;
+	const tc = tri.c;
+
+	let i0 = i;
+	let i1 = i + 1;
+	let i2 = i + 2;
+	if ( index ) {
+
+		i0 = index.getX( i );
+		i1 = index.getX( i + 1 );
+		i2 = index.getX( i + 2 );
+
+	}
+
+	ta.x = pos.getX( i0 );
+	ta.y = pos.getY( i0 );
+	ta.z = pos.getZ( i0 );
+
+	tb.x = pos.getX( i1 );
+	tb.y = pos.getY( i1 );
+	tb.z = pos.getZ( i1 );
+
+	tc.x = pos.getX( i2 );
+	tc.y = pos.getY( i2 );
+	tc.z = pos.getZ( i2 );
 
 }
