+++ src/core/castFunctions.js
