+++ src/math/MathUtilities.js
@@ -1,4 +1,5 @@
 import { Vector3, Vector2, Plane, Line3 } from 'three';
+
 export const closestPointLineToLine = ( function () {
 
 	// https://github.com/juj/MathGeoLib/blob/master/src/Geometry/Line.cpp#L56
