+++ src/math/SeparatingAxisTriangle.js
